DFE-905 - broke Edit Release sections into individual pages
