+++ src/explore-education-statistics-admin/src/pages/edit-release/EditReleaseDataPage.tsx
@@ -1,0 +1,43 @@
+import React, { useEffect, useState } from 'react';
+import { RouteComponentProps } from 'react-router';
+import DummyPublicationsData from '@admin/pages/DummyPublicationsData';
+import EditReleasePageTemplate, {
+  ReleaseSection,
+} from '@admin/pages/edit-release/components/EditReleasePageTemplate';
+import { Release } from '../../services/publicationService';
+
+interface MatchProps {
+  releaseId: string;
+}
+
+const EditReleaseDataPage = ({ match }: RouteComponentProps<MatchProps>) => {
+  const { releaseId } = match.params;
+
+  const [release, setRelease] = useState<Release>();
+
+  const [publicationTitle, setPublicationTitle] = useState('');
+
+  useEffect(() => {
+    const selectedRelease = DummyPublicationsData.getReleaseById(releaseId);
+
+    const owningPublication = DummyPublicationsData.getOwningPublicationForRelease(
+      selectedRelease,
+    );
+
+    setRelease(selectedRelease);
+
+    setPublicationTitle(owningPublication ? owningPublication.title : '');
+  }, [releaseId]);
+
+  return (
+    <EditReleasePageTemplate
+      publicationTitle={publicationTitle}
+      releaseId={releaseId}
+      selectedSection={ReleaseSection.AddEditData}
+    >
+      {release && <h2 className="govuk-heading-m">Add / edit data</h2>}
+    </EditReleasePageTemplate>
+  );
+};
+
+export default EditReleaseDataPage;
